Add files via upload
The objective of this project is to provide a comprehensive overview of the development and implementation of a Music Recommendation System using Machine Learning. It aims to describe the design, data collection, algorithms, and evaluation methods used in creating an effective music recommendation system. It also explores the underlying technologies and techniques involved in building a personalized music recommendation system that enhances user experience in the digital music domain.